dockerize app
