+++ NBA/templates/player-ranking.html
@@ -1,0 +1,28 @@
+<html>
+<head>
+    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@4.3.1/dist/css/bootstrap.min.css" integrity="sha384-ggOyR0iXCbMQv3Xipma34MD+dH/1fQ784/j6cY/iJTQUOhcWr7x9JvoRxT2MZw1T" crossorigin="anonymous">
+
+</head>
+<body>
+    <table class="table">
+			<thead>
+				<tr>
+					<th scope="col">Rrank</th>
+					{% for key,val in object_list.0.items %}
+					<th scope="col">{{key}}</th>
+					{% endfor %}
+				</tr>
+			</thead>
+			<tbody>
+					{% for player in object_list %}
+					<tr>
+						<th scope="row">{{forloop.counter}}</th>
+					{% for key,val in player.items%}
+					<td>{{val}}</td>
+					{% endfor %}
+					</tr>
+					{% endfor %}
+			</tbody>
+    </table>
+</body>
+</html>
