+++ docker-compose.yml
@@ -6,11 +6,9 @@ services:
     container_name: django_nba
     command: bash -c "
         python manage.py makemigrations 
-        && python manage.py migrate 
-        && python manage.py runscript initial_compute
+        && python manage.py migrate
         && python manage.py runscript update_setup
         && python manage.py runserver 0.0.0.0:8000"
-
     ports:
       - "8000:8000"
     environment:
@@ -21,9 +19,21 @@ services:
       - type: bind
         source: ./NBA
         target: /NBA
+      - type: bind
+        source: ./scripts
+        target: /scripts
     depends_on:
       - pgdb
       - redis
+  beat:
+    build: .
+    command: bash -c " python manage.py runscript update_setup
+        && celery -A NBA beat -l info --scheduler django_celery_beat.schedulers:DatabaseScheduler"
+    container_name: beat_nba
+    depends_on:
+      - django
+      - redis
+      - celery
   celery:
     build: .
     command: bash -c "python manage.py makemigrations 
@@ -33,6 +43,9 @@ services:
       - type: bind
         source: ./NBA
         target: /NBA
+      - type: bind
+        source: ./scripts
+        target: /scripts
     depends_on:
         - django
         - redis
