+++ requirements.txt
@@ -28,6 +28,7 @@ python-dateutil==2.8.2
 pytz==2023.3
 redis==4.6.0
 requests==2.31.0
+schedule==1.2.0
 six==1.16.0
 sqlparse==0.4.4
 typing_extensions==4.6.3
